Add basic filter cycling functionality
