+++ README.md
@@ -27,7 +27,8 @@ You can check the working version from [here](https://excalith-start-page.vercel
 - Search websites with custom commands. For example, type `s some weird bug` to search StackOverflow for `some weird bug`
 - Wallpaper support through URL with blur and fade effects
 - Customizable Fetch UI for fetching browser and system data, including custom image support
-- Autosuggest and <kbd>TAB</kbd> Autocomplete just like `zsh` and `fish`
+- Autosuggest and Autocomplete support just like `zsh` and `fish`
+- Cycle through filtered links back and forth
 - Multiple theme support (check all [available themes](./public/themes/))
 - Built-in configuration editor to easily edit and save your configuration
 
@@ -48,6 +49,7 @@ Please refer to [configuration](https://github.com/excalith/excalith-start-page/
 ### Key Bindings
 
 - Use <kbd>→</kbd> to auto-complete the suggestion
+- Cycle filtered links using <kbd>TAB</kbd> and <kbd>SHIFT</kbd> + <kbd> TAB</kbd>
 - Clear the prompt quickly with <kbd>CTRL</kbd> + <kbd>C</kbd>
 - Close window with <kbd>ESC</kbd>
