+++ src/components/List.js
@@ -2,6 +2,7 @@ import React, { use, useEffect, useState } from "react"
 import Link from "@/components/Link"
 import Search from "@/components/Search"
 import { useSettings } from "@/context/settings"
+import { openLink } from "@/utils/openLink"
 
 const Section = ({ section, filter, selection }) => {
 	const alignment = section.align || "left"
@@ -37,11 +38,19 @@ const List = () => {
 	const [selection, setSelection] = useState("")
 
 	useEffect(() => {
+		if (filteredItems.length <= 1) {
+			setSelection("")
+		}
+
 		const handleKeyDown = (event) => {
 			if (command === "") return
 
+			// Submit
+			if (event.key === "Enter") {
+				openFilteredLinks()
+			}
 			// Previous Selection
-			if (event.shiftKey && event.key === "Tab") {
+			else if (event.shiftKey && event.key === "Tab") {
 				let idx = -1
 				if (selection && selection !== "")
 					idx = filteredItems.indexOf(selection.toLowerCase())
@@ -75,6 +84,50 @@ const List = () => {
 		const filtered = items.filter((item) => item.startsWith(lowerCommand))
 		setFilteredItems(filtered)
 	}, [items, command])
+
+	const openFilteredLinks = () => {
+		console.log("selection: " + selection)
+		console.log(filteredItems)
+
+		if (selection !== "") {
+			settings.sections.list.map((section) => {
+				{
+					section.links.map((link) => {
+						{
+							if (link.name.toLowerCase() === selection) {
+								openLink(link.url)
+								return
+							}
+						}
+					})
+				}
+			})
+		} else {
+			let filteredUrls = []
+			settings.sections.list.map((section) => {
+				{
+					section.links.map((link) => {
+						{
+							if (link.name.toLowerCase().includes(command)) {
+								filteredUrls.push(link.url)
+							}
+						}
+					})
+				}
+			})
+
+			let filterCount = filteredUrls.length
+			if (filterCount === 0) {
+				const defaultSerachEngine = settings.search.default
+				const target = settings.search.target
+				openLink(defaultSerachEngine + command, target)
+			} else {
+				filteredUrls.map((url, index) => {
+					openLink(url, index === filterCount - 1 ? "_self" : "_blank")
+				})
+			}
+		}
+	}
 
 	const handleCommandChange = (str) => {
 		setCommand(str)
