+++ src/components/Search.js
@@ -16,9 +16,7 @@ const Search = ({ prompt, commandChange }) => {
 
 	useEffect(() => {
 		const handleKeyDown = (event) => {
-			if (event.key === "Enter") {
-				RunCommand(input.current.value, settings)
-			} else if (settings.prompt.ctrlC) {
+			if (settings.prompt.ctrlC) {
 				if ((event.metaKey || event.ctrlKey) && event.code === "KeyC") {
 					input.current.value = ""
 					setSuggestion("")
